rename package, add edittext for firstname and lastname
